requirements.txt
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-storages==1.7.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
gunicorn==19.9.0
idna==2.8
Pillow==6.2.1
protobuf==3.9.2
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.2
razorpay==1.2.0
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
